+++ CHANGELOG.md
@@ -5,7 +5,7 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3.1] - 2023-04-17
 
 ### Added
 
@@ -37,7 +37,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Move from giganto
 
-[Unreleased]: https://github.com/aicers/giganto-client/compare/0.3.0...main
+[0.3.1]: https://github.com/aicers/giganto-client/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/aicers/giganto-client/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/aicers/giganto-client/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/aicers/giganto-client/tree/0.1.0